+++ src/main/java/Application.java
@@ -1,15 +1,15 @@
 import blackjack.BlackjackController;
 import blackjack.config.GameConfig;
-import blackjack.view.ConsoleInputView;
-import blackjack.view.ConsoleOutputView;
+import blackjack.view.InputView;
+import blackjack.view.OutputView;
 
 public class Application {
 
     public static void main(String[] args) {
 
         GameConfig gameConfig = new GameConfig(
-                new ConsoleInputView(),
-                new ConsoleOutputView()
+                new InputView(),
+                new OutputView()
         );
         BlackjackController blackJackController = new BlackjackController(gameConfig);
         blackJackController.run();
